Reorganizing new test classes
Breaking the new abstract parameterizations and percent codec test
implementations into sub-packages within test scope.  Updating
dependencies and references to resolve compilation issues.
