+++ src/test/java/org/owasp/esapi/codecs/percent/PercentCodecStringTest.java
@@ -1,4 +1,4 @@
-package org.owasp.esapi.codecs;
+package org.owasp.esapi.codecs.percent;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -6,6 +6,7 @@ import java.util.List;
 
 import org.junit.runners.Parameterized.Parameters;
 import org.owasp.esapi.codecs.PercentCodec;
+import org.owasp.esapi.codecs.abstraction.AbstractCodecStringTest;
 
 /**
  *  Codec validation focused on the PercentCodec String-based api.
